Added payment feature to help convert currencies if that is needed. Fixed the radius search in car_location

+++ car_location/common/common.go
@@ -14,7 +14,7 @@ type CarLocation struct{
 
 type Location struct{
 	Coords 	string `json:"location"`
-	Radius  int    `json:"radius"`
+	Radius  float32`json:"radius"`
 }
